MLNanoShaper

Documentation for MLNanoShaper.

MLNanoShaper.AccumulatorLogger

MLNanoShaper.AuxiliaryParameters

MLNanoShaper.TrainingData

MLNanoShaper.TrainingParameters

MLNanoShaperRunner.AnnotedKDTree

MLNanoShaperRunner.ConcatenatedBatch

MLNanoShaperRunner.ModelInput

MLNanoShaperRunner.Option

MLNanoShaper.categorical_loss

MLNanoShaper.comonicon_install

MLNanoShaper.comonicon_install_path

MLNanoShaper.continus_loss

MLNanoShaper.evaluate_model

MLNanoShaper.generate_data

MLNanoShaper.generate_data_points

MLNanoShaper.implicit_surface

MLNanoShaper.load_data_pdb

MLNanoShaper.load_data_pqr

MLNanoShaper.train

MLNanoShaper.train

MLNanoShaper.train

MLNanoShaperRunner.anakin

MLNanoShaperRunner.distance

MLNanoShaperRunner.eval_model

MLNanoShaperRunner.load_atoms

MLNanoShaperRunner.load_model

MLNanoShaper.AccumulatorLogger — Type

accumulator(processing,logger)

A processing logger that transform logger on multiple batches Ca be used to liss numerical data, for logging to TensorBoardLogger.

MLNanoShaper.AuxiliaryParameters — Type

AuxiliaryParameters

The variables that do not influence the outome of the training run. This include the nb_epoch.

MLNanoShaper.TrainingData — Type

Training information used in model training.

Fields

atoms: the set of atoms used as model input

skin : the Surface generated by Nanoshaper

MLNanoShaper.TrainingParameters — Type

TrainingParameters

The training parameters used in the model training. Default values are in the param file. The training is deterministric. Theses values are hased to determine a training run

MLNanoShaper.categorical_loss — Method

categorical_loss(model, ps, st, (; point, atoms, d_real))

The loss function used by in training. Return the KL divergence between true probability and empirical probability Return the error with the espected distance as a metric.

MLNanoShaper.comonicon_install — Method

comonicon_install(;kwargs...)

Install the CLI manually. This will use the default configuration in Comonicon.toml, if it exists. For more detailed reference, please refer to Comonicon documentation.

MLNanoShaper.comonicon_install_path — Method

comonicon_install_path(;[yes=false])

Install the PATH and FPATH to your shell configuration file. You can use comonicon_install_path(;yes=true) to skip interactive prompt. For more detailed reference, please refer to Comonicon documentation.

MLNanoShaper.continus_loss — Method

continus_loss(model, ps, st, (; point, atoms, d_real))

The loss function used by in training. compare the predicted (square) distance with $\frac{1 + anh(d)}{2}$ Return the error with the espected distance as a metric.

MLNanoShaper.evaluate_model — Method

evaluate_model(
    model::Lux.StatefulLuxLayer, x::Point3f, atoms::AnnotedKDTree; cutoff_radius, default_value = -0.0f0)

evaluate the model on a single point.
This function handle the logic in case the point is too far from the atoms. In this case default_value is returned and the model is not run.

MLNanoShaper.generate_data — Method

generate_data()

generate data from the parameters files in param/ by downloading the pdb files and running Nanoshaper.

MLNanoShaper.generate_data_points — Method

generate_data_points(
    preprocessing::Lux.AbstractExplicitLayer, points::AbstractVector{<:Point3},
    (; atoms, skin)::TreeTrainingData{Float32}, (; ref_distance)::TrainingParameters)

generate the data_points for a set of positions points on one protein.

MLNanoShaper.implicit_surface — Method

implicit_surface(atoms::AnnotedKDTree{Sphere{T}, :center, Point3{T}},
    model::Lux.StatefulLuxLayer, (;
        cutoff_radius, step)) where {T}

Create a mesh form the isosurface of function `pos -> model(atoms,pos)` using marching cubes algorithm and using step size `step`.

MLNanoShaper.load_data_pdb — Method

load_data_pdb(T, name::String)

Load a TrainingData{T} from current directory. You should have a pdb and an off file with name name in current directory.

MLNanoShaper.load_data_pqr — Method

load_data_pqr(T, name::String)

Load a TrainingData{T} from current directory. You should have a pdb and an off file with name name in current directory.

MLNanoShaper.train — Function

train [options] [flags]

Train a model.

Intro

Parameters are specified in the param/param.toml file. The folowing parameters can be overided.

Options

--nb-epoch, -e <Int>: the number of epoch to compute.

--model, -m <String>: the model name. Can be anakin.

--nb-data-points, -d <Int>: the number of proteins in the dataset to use

--name, -n <String>: name of the training run

--cutoff-radius, -c <Float32>: the cutoff_radius used in training

--ref-distance, -d <Float32>: the reference distane (in A) used to rescale distance to surface in loss

Flags

--gpu, -g: should we do the training on the gpu

--categorical, -c: should we use the categorical version of the model

MLNanoShaper.train — Method

train(training_parameters::TrainingParameters, directories::AuxiliaryParameters)

train the model given TrainingParameters and AuxiliaryParameters.

MLNanoShaper.train — Method

train((train_data,test_data),training_states; nb_epoch)

train the model on the data with nb_epoch

MLNanoShaperRunner.Option — Type

state

The global state manipulated by the c interface. To use, you must first load the weights using load_weights and the input atoms using load_atoms. Then you can call eval_model to get the field on a certain point.

MLNanoShaperRunner.AnnotedKDTree — Type

AnnotedKDTree(data::StructVector,property::StaticSymbol)

Fields

data::StructVector

tree::KDTree

MLNanoShaperRunner.ConcatenatedBatch — Type

ConcatenatedBatch

Represent a vector of arrays of sizes (a..., bn) where bn is the variable dimension of the batch. You can access view of individual arrays with get_slice.

MLNanoShaperRunner.ModelInput — Type

ModelInput

input of the model

Fields

point::Point3, the position of the input

atoms::StructVector{Sphere}, the atoms in the neighboord

MLNanoShaperRunner.anakin — Method

anakin()

A model inspired by the anakin paper. The protein is decomposed in pairs of atom near the interest point. Each pair of atom is turned into a Preprocessed data and is then fed to the Encoding layer.

MLNanoShaperRunner.distance — Method

distance(x::GeometryBasics.Mesh, y::KDTree)

Return the Hausdorff distance betwen the mesh coordinates

MLNanoShaperRunner.eval_model — Function

eval_model(x::Float32,y::Float32,z::Float32)::Float32

evaluate the model at coordinates x y z.

MLNanoShaperRunner.load_atoms — Function

load_atoms(start::Ptr{CSphere},length::Cint)::Cint

Load the atoms into the julia model. Start is a pointer to the start of the array of CSphere and length is the length of the array

Return an error status:

0: OK

1: data could not be read

2: unknow error

MLNanoShaperRunner.load_model — Function

load_model(path::String)::Cint

Load the model from a MLNanoShaperRunner.SerializedModel serialized state at absolute path path.

Return an error status:

0: OK

1: file not found

2: file could not be deserialized properly

3: unknow error